Moving the main migration to the right directory
